requirements.txt
carla==0.9.12
carla.egg==info
gym==0.18.0
networkx==2.6.3
numpy==1.21.2
opencv_python==4.5.3.56
Pillow==8.4.0
psutil==5.8.0
pygame==2.0.1
PyYAML==6.0
tensorboard==2.6.0
tqdm==4.62.2
webdataset==0.1.62
